+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Return searchInput property of HubMainScreen state (gorarakelyan)
+
 ## 1.1.7 Oct 5 2020
 - Update web page title (gorarakelyan)